refactor: convert feedback form to server action
