+++ package.json
@@ -55,8 +55,8 @@
     "@types/lodash": "^4.14.195",
     "@types/module-alias": "^2.0.1",
     "@types/node": "20.4.5",
-    "@types/react": "18.2.17",
-    "@types/react-dom": "18.2.7",
+    "@types/react": "^18.2.73",
+    "@types/react-dom": "^18.2.22",
     "@types/testing-library__jest-dom": "^5.14.9",
     "@types/whatwg-fetch": "^0.0.33",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
